chore: position type

+++ controllers/app/api/v1/app_types.go
@@ -58,7 +58,7 @@ type AppSpec struct {
 	AppMeta `json:",inline"`
 
 	//+kubebuilder:validation:Optional
-	Position int32 `json:"position,omitempty"`
+	Position float64 `json:"position,omitempty"`
 
 	//+kubebuilder:validation:Optional
 	I18N *map[string]AppMeta `json:"i18n,omitempty"`
